requirements.txt
click
requests
beautifulsoup4
youtube-transcript-api
feedparser
pyyaml
openai
anthropic
python-dotenv
pydantic
